Screen characters list

+++ app/src/main/java/com/lidorttol/rickandmorty/data/local/dbo/character/CharacterEpisodeCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["characterId", "episodeId"])
 data class CharacterEpisodeCrossRef(
-    val characterId: Int,
-    val episodeId: Int
+    val characterId: Long,
+    val episodeId: Long
 )
